--- example/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "example",
-  "version": "1.0.0",
-  "main": "index.tsx",
-  "license": "MIT",
-  "source": "./index.tsx",
-  "scripts": {
-    "start": "NODE_ENV=development yarn run tailwindcss && start-storybook -p 6006",
-    "build": "NODE_ENV=production yarn run tailwindcss:prod && build-storybook",
-    "tailwindcss": "tailwindcss build -o ./tailwind.css",
-    "tailwindcss:prod": "tailwindcss build -o ./tailwind.css"
-  },
-  "dependencies": {
-    "react": "^16.13.1",
-    "react-app-polyfill": "^1.0.0",
-    "react-dom": "^16.13.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.11.4",
-    "@storybook/addon-actions": "^6.0.16",
-    "@storybook/addon-essentials": "^6.0.16",
-    "@storybook/addon-links": "^6.0.16",
-    "@storybook/react": "^6.0.16",
-    "@tailwindcss/ui": "^0.5.0",
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.8.4",
-    "babel-loader": "^8.1.0",
-    "react-is": "^16.13.1",
-    "tailwindcss": "^1.6.2",
-    "typescript": "3.4.5"
-  }
-}
