--- example/vercel.json
@@ -1,6 +1,0 @@
-{
-  "name": "react-listbox",
-  "github": {
-    "silent": true
-  }
-}
