+++ package.json
@@ -15,9 +15,13 @@
   "scripts": {
     "start": "tsdx watch",
     "build": "tsdx build",
+    "storybook": "NODE_ENV=development yarn run tailwindcss && start-storybook -p 6006",
+    "build-storybook": "NODE_ENV=production yarn run tailwindcss:prod && build-storybook",
     "test": "tsdx test --passWithNoTests",
     "lint": "tsdx lint src test example",
-    "prepare": "tsdx build"
+    "prepare": "tsdx build",
+    "tailwindcss": "tailwindcss build -o ./stories/tailwind.css",
+    "tailwindcss:prod": "tailwindcss build -o ./stories/tailwind.css"
   },
   "peerDependencies": {
     "react": ">=16"
@@ -29,10 +33,16 @@
   },
   "prettier": {},
   "devDependencies": {
+    "@storybook/addon-actions": "^6.0.16",
+    "@storybook/addon-essentials": "^6.0.16",
+    "@storybook/addon-links": "^6.0.16",
+    "@storybook/react": "^6.0.16",
+    "@tailwindcss/ui": "^0.5.0",
     "@types/react": "^16.9.46",
     "@types/react-dom": "^16.9.8",
     "@typescript-eslint/eslint-plugin": "^3.9.0",
     "@typescript-eslint/parser": "^3.9.0",
+    "concurrent": "^0.3.2",
     "eslint": "^7.7.0",
     "eslint-config-react-app": "^5.2.1",
     "eslint-config-standard": "^14.1.1",
@@ -49,6 +59,7 @@
     "husky": "^4.2.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "tailwindcss": "^1.7.6",
     "tsdx": "^0.13.2",
     "tslib": "^2.0.1",
     "typescript": "^3.9.7"
