+++ tailwind.config.js
